python3: remove casts from "int" to "long"
There has not been any distinction at the python code level between `long`
and `int` since 2.2.

`2to3 -f long` would replace them with casts to `int` instead, which would
just be a no-op.

+++ xcp/net/mac.py
@@ -88,8 +88,8 @@ class MAC(object):
             raise ValueError("Expected 6 octets, got %d" % len(octets))
 
         self.octets = [ int(i, 16) for i in octets ]
-        self.integer = long(sum(t[0] << t[1] for t in
-                                zip(self.octets, range(40, -1, -8))))
+        self.integer = sum(t[0] << t[1] for t in
+                           zip(self.octets, range(40, -1, -8)))
 
     def _set_from_str_quads(self, quads):
         """Private helper"""
@@ -100,8 +100,8 @@ class MAC(object):
         for quad in ( int(i, 16) for i in quads ):
             self.octets.extend([(quad >> 8) & 0xff, quad & 0xff])
 
-        self.integer = long(sum(t[0] << t[1] for t in
-                                zip(self.octets, range(40, -1, -8))))
+        self.integer = sum(t[0] << t[1] for t in
+                           zip(self.octets, range(40, -1, -8)))
 
     def is_unicast(self):
         """is this a unicast address?"""
